moved data builder class to data package and updated testng xml to run all the tests
