+++ src/test/java/io/github/mfaisalkhatri/RestfulBookerTests.java
@@ -3,7 +3,7 @@ package io.github.mfaisalkhatri;
 import io.github.mfaisalkhatri.data.CreateBookingData;
 import org.testng.annotations.Test;
 
-import static io.github.mfaisalkhatri.CreateBookingDataBuilder.getBookingData;
+import static io.github.mfaisalkhatri.data.CreateBookingDataBuilder.getBookingData;
 import static io.restassured.RestAssured.given;
 import static org.hamcrest.Matchers.*;
